rename EmailVerificationView to VerifyEmailView
