+++ english_diary/profiles/urls.py
@@ -4,7 +4,7 @@ from profiles.views import *
 
 
 urlpatterns = [
-    url(r'^verify/(?P<verification_key>\w+)/$', EmailVerificationView.as_view(), name="email_verification"),
+    url(r'^verify/(?P<verification_key>\w+)/$', VerifyEmailView.as_view(), name="verify_email"),
     url(r'^verify/(?P<verification_key>\w+)/expire/$', KeyExpiresView.as_view(), name="key_expires"),
     url(r'^renew/$', RenewKeyView.as_view(), name="renew_key"),
 ]
