+++ english_diary/profiles/views/__init__.py
@@ -1,3 +1,3 @@
-from .email_verification import EmailVerificationView
+from .verify_email import VerifyEmailView
 from .key_expires import KeyExpiresView
 from .renew_key import RenewKeyView
