+++ english_diary/profiles/views/verify_email.py
